backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.12.1
geoalchemy2==0.14.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Redis
redis==5.0.1
hiredis==2.2.0

# AI/ML
openai==1.3.5

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
structlog==23.2.0
email-validator==2.1.0
phonenumbers==8.13.26

# Google Maps
googlemaps==4.10.0

# Date/Time
python-dateutil==2.8.2

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

